requirements.txt
pydub==0.25.1
pillow==10.0.0
pyinstaller==6.1.0
